pyproject.toml
[project]
name = "fastmvc_dashboards"
version = "0.3.0"
description = "Dashboards for FastMVC: embed signing, revocation, theme params, Metabase/Grafana/Looker/PowerBI stubs, routers."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "httpx>=0.24.0",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0", "httpx>=0.24.0", "PyJWT>=2.0.0"]
metabase = ["PyJWT>=2.0.0"]

[project.urls]
Homepage = "https://github.com/your-org/fastmvc_dashboards"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src/fastmvc_dashboards"]
branch = true
# Integration-heavy routers (host app imports) are not covered by unit tests; measure embed + providers + layout.
omit = [
    "*/_optional_import.py",
    "*/api_dashboard/router.py",
    "*/api_dashboard/registry.py",
    "*/health/dashboard.py",
    "*/queues_dashboard/router.py",
    "*/secrets_dashboard/router.py",
    "*/tenants_dashboard/router.py",
    "*/workflows_dashboard/router.py",
    "*/workflows/*.py",
    "*/secrets/*.py",
    "*/router.py",
    "src/fastmvc_dashboards/__init__.py",
    "*/queues_dashboard/__init__.py",
    "*/secrets_dashboard/__init__.py",
    "*/tenants_dashboard/__init__.py",
    "*/workflows_dashboard/__init__.py",
    "*/health/__init__.py",
]

[tool.coverage.report]
fail_under = 90
show_missing = true
exclude_lines = ["pragma: no cover"]
omit = ["*/tests/*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
